pkg/deck: persist a deck of cards
This commit introduces the interface for persistent deck state, we are
also including an in-memory implementation.

Since implementations of the interface will likely include heavier IO
operations or network calls, a context had been added to provide
cancellation.

A more proper solution would require a more persistent store
(e.g., postgresql, redis) but for the simplicity of the project
an in-memory implementation is sufficient.