go.mod: go get github.com/chn555/schemas
In this repo we are working schema-driven, with our schema being
protobuf files hosted on a separate repo. The code generated by the
schema is required for use to implement the services defined in the
schema, and is contained in the schemas repo.

I am not including uuid package in the commit header as shown in
https://github.com/OneLayerHQ/atmosphere/pull/444/commits/71b0e26ba604dec977ec219c5321b0f9ca7e5990